requirements.txt
datasets==4.0.0
faiss-cpu==1.8.0.post1
flash_attn==2.6.3
jsonlines==4.0.0
networkx==3.2.1
numpy==1.23.5
openai==1.46.0
opencv-python==4.10.0.84
pandas==2.2.2
peft==0.17.0
rouge==1.0.1
tensorboard==2.20.0
tokenizers==0.21.4
torch==2.4.0
tqdm==4.66.4
transformers==4.55.0
trl==0.21.0
vllm==0.5.4
vllm-flash-attn==2.6.1